Add platform-specific configuration and shared constants for ASoC SAST Bitbucket Pipe
- Created common/constants.py to define centralized constants for the ASoC SAST pipeline.
- Added platform_config.py for Linux with specific constants and download endpoint.
- Added platform_config.py for Windows with specific constants and download endpoint.
- Introduced versioning and various API endpoints for better integration.
